refactor: improve folder structure, remove duplicate code, minor ui adjustment
