+++ public/assets/images/icons/projectV2/FieldReviewedIcon.tsx
