+++ public/assets/images/icons/projectV2/NewInfoIcon.tsx
