+++ public/assets/images/icons/projectV2/OffSiteReviewedIcon.tsx
