+++ public/assets/images/icons/projectV2/TopProjectIcon.tsx
