+++ public/static/locales/en/projectDetails.json
@@ -4,7 +4,7 @@
     "diveIntoTheProject": "Dive into the project",
     "firePopupText": "We only display fires that occurred within the past 12 months.",
     "fieldReviewed": "Field Reviewed",
-    "offSiteReviewed": "Off-Site reviewed",
+    "offSiteReviewed": "Off-Site Reviewed",
     "report": "Report",
     "seeLess": "see less",
     "seeMore": "see more",
