+++ src/features/projects/screens/SingleProjectDetails.tsx
@@ -11,7 +11,7 @@ import { useLocale, useTranslations } from 'next-intl';
 import CancelIcon from '../../../../public/assets/images/icons/CancelIcon';
 import ExpandIcon from '../../../../public/assets/images/icons/ExpandIcon';
 import ProjectInfo from '../components/projectDetails/ProjectInfo';
-import ProjectSnippet from '../components/ProjectSnippet';
+import ProjectSnippet from '../../projectsV2/components/ProjectSnippet';
 import SitesDropdown from '../components/maps/SitesDropdown';
 import { useProjectProps } from '../../common/Layout/ProjectPropsContext';
 import ProjectTabs from '../components/maps/ProjectTabs';
