+++ src/features/projects/styles/Projects.scss
@@ -1,5 +1,5 @@
 @import './../../../theme/theme';
-@import './ProjectSnippet.module.scss';
+@import '../../projectsV2/styles/ProjectSnippet.module.scss';
 @import './ProjectDetails.module.scss';
 
 .container,
