+++ src/features/projectsV2/components/microComponents/ProjectTypeIcon.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import Agroforestry from '../../../../public/assets/images/icons/project/Agroforestry';
-import Mangroves from '../../../../public/assets/images/icons/project/Mangroves';
-import NaturalRegeneration from '../../../../public/assets/images/icons/project/NaturalRegeneration';
-import UrbanRestoration from '../../../../public/assets/images/icons/project/UrbanRestoration';
-import Conservation from '../../../../public/assets/images/icons/project/Conservation';
-import ManagedRegeneration from '../../../../public/assets/images/icons/project/ManagedRegeneration';
-import TreePlanting from '../../../../public/assets/images/icons/project/TreePlanting';
+import Agroforestry from '../../../../../public/assets/images/icons/project/Agroforestry';
+import Mangroves from '../../../../../public/assets/images/icons/project/Mangroves';
+import NaturalRegeneration from '../../../../../public/assets/images/icons/project/NaturalRegeneration';
+import UrbanRestoration from '../../../../../public/assets/images/icons/project/UrbanRestoration';
+import Conservation from '../../../../../public/assets/images/icons/project/Conservation';
+import ManagedRegeneration from '../../../../../public/assets/images/icons/project/ManagedRegeneration';
+import TreePlanting from '../../../../../public/assets/images/icons/project/TreePlanting';
 interface Props {
   projectType: string | null;
 }
