+++ src/features/projectsV2/styles/Badge.module.scss
@@ -1,17 +1,22 @@
-@import '../../theme/theme';
+@import '../../../theme/theme';
 
 .projectBadge {
   display: flex;
   height: 27px;
   gap: 3.87px;
   border-radius: 0px 16px;
-  border: 3px solid $backgroundColor;
+  border: 2px solid $backgroundColor;
   align-items: center;
   padding: 0px 17px;
   white-space: nowrap;
   font-weight: 700;
   background: $dark;
   width: fit-content;
+  cursor: pointer;
+  position: absolute;
+  z-index: 1;
+  right: -2px;
+  top: -2px;
 }
 .badgeIcon {
   display: flex;
@@ -29,9 +34,7 @@
   max-width: 300px;
 }
 
-.projectSnippetTooltip {
-  position: absolute;
-  cursor: pointer;
-  top: -3px;
-  right: -3px;
+.standardsLink {
+  font-weight: 400;
+  color: $primaryColor;
 }
