+++ src/theme/_colors.scss
@@ -32,6 +32,7 @@ $certificationLinkColorNew: var(--certification-link-color-new);
 $secondaryDividerColorNew: var(--secondary-divider-color-new);
 $grayFontColorNew: var(--gray-font-color-new);
 $lightGrayBackgroundColorNew: var(--light-gray-background-color-new);
+$charcoalGray: var(--charcoal-gray);
 $boldFontColorNew: var(--bold-font-color-new);
 $selectedMenuItemColorNew: var(--selected-menu-item-color-new);
 $dividerColorNew: var(--divider-color-new);
